==> Tbrain/Tbrian_BryanChi_solution-master/2_basic_feature1_stat.ipynb <==
## this notebook
#1. add day_num (ith day i=1...92)
#2. create day_cnt df, first column is fild_id feature is 0 , 1 , ... 92 indicate how many data corresponding to that file in that day_num
#3. create cust_cnt, prod_cnt, count how many customer_id, product_id, corresponding to that file
#4. create week_cnt, hour_cnt, min_cnt, sec_cnt count how many data corresponding to that time
#5. create prod_id_cnt count how many data corresponding to that product_id (divide by prod_sum)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import gc

%matplotlib inline

# day_cnt

#let's create daily dictionary
daily = list(np.arange(31)+1)
month = ['03','04','05']
dailydict = dict()
j=300
for i in daily:
    j=j+1
    dailydict[j]=i
j=400
for i in daily[:-1]:
    j=j+1
    dailydict[j]=i+31
j=500
for i in daily:
    j=j+1
    dailydict[j]=i+61

march = pd.read_hdf('march.h5', 'df')
march['day_num']=march['tm_mon']*100 + march['tm_day']
march['day_num'] = march['day_num'].map(dailydict)
march.to_hdf('march.h5','df')


april = pd.read_hdf('april.h5', 'df')
april['day_num']=april['tm_mon']*100 + april['tm_day']
april['day_num'] = april['day_num'].map(dailydict)
april.to_hdf('april.h5','df')


may = pd.read_hdf('may.h5', 'df')
may['day_num']=may['tm_mon']*100 + may['tm_day']
may['day_num'] = may['day_num'].map(dailydict)
may.to_hdf('may.h5','df')


march = pd.read_hdf('march.h5', 'df')
march_cnt = march[['file_id','day_num','timestamp']].groupby(['file_id','day_num']).count().reset_index()
march_cnt.columns = ['file_id','day_num','frequency']
march_cnt = march_cnt.pivot_table(values='frequency',index=['file_id'],columns='day_num',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_cnt = april[['file_id','day_num','timestamp']].groupby(['file_id','day_num']).count().reset_index()
april_cnt.columns = ['file_id','day_num','frequency']
april_cnt = april_cnt.pivot_table(values='frequency',index=['file_id'],columns='day_num',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_cnt = may[['file_id','day_num','timestamp']].groupby(['file_id','day_num']).count().reset_index()
may_cnt.columns = ['file_id','day_num','frequency']
may_cnt = may_cnt.pivot_table(values='frequency',index=['file_id'],columns='day_num',aggfunc='sum').reset_index()

del may
gc.collect();

march_cnt.head(3)

april_cnt.head(3)

may_cnt.head(3)

data = pd.merge(march_cnt,april_cnt,how='outer',on='file_id')
data = pd.merge(data,may_cnt,how='outer',on='file_id')

data.head(3)

#data.to_csv('day_cnt.csv',index=False)

# prod_cnt cust_cnt

march = pd.read_hdf('march.h5', 'df')
march = march[['file_id','customer_id','timestamp']]
march = march.groupby(['file_id','customer_id']).count().reset_index()

april = pd.read_hdf('april.h5', 'df')
april = april[['file_id','customer_id','timestamp']]
april = april.groupby(['file_id','customer_id']).count().reset_index()

may = pd.read_hdf('may.h5', 'df')
may = may[['file_id','customer_id','timestamp']]
may = may.groupby(['file_id','customer_id']).count().reset_index()

march.head(3)

data = pd.concat([march,april],axis=0)
data = pd.concat([data,may],axis=0)

data.head(3)

data=data.groupby(['file_id','customer_id']).count().reset_index()
data = data.groupby('file_id').count().reset_index()

data.head(3)

cust_prod = data[['file_id','timestamp']]
cust_prod.columns = ['file_id','cust_cnt']
cust_prod.to_csv('cust_prod.csv',index=False)

march = pd.read_hdf('march.h5', 'df')
march = march[['file_id','product_id','timestamp']]
march = march.groupby(['file_id','product_id']).count().reset_index()

april = pd.read_hdf('april.h5', 'df')
april = april[['file_id','product_id','timestamp']]
april = april.groupby(['file_id','product_id']).count().reset_index()

may = pd.read_hdf('may.h5', 'df')
may = may[['file_id','product_id','timestamp']]
may = may.groupby(['file_id','product_id']).count().reset_index()

data = pd.concat([march,april],axis=0)
data = pd.concat([data,may],axis=0)
data=data.groupby(['file_id','product_id']).count().reset_index()
data = data.groupby('file_id').count().reset_index()

data.head(3)

data = data[['file_id','timestamp']]
data.columns = ['file_id','prod_cnt']
cust_prod = pd.merge(cust_prod,data,how='left',on='file_id')

cust_prod.head(3)

cust_prod.to_csv('cust_prod.csv',index=False)

# week_cnt

march = pd.read_hdf('march.h5', 'df')
march_cnt = march[['file_id','tm_wday','timestamp']].groupby(['file_id','tm_wday']).count().reset_index()
march_cnt.columns = ['file_id','tm_wday','frequency']
march_cnt = march_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_wday',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_cnt = april[['file_id','tm_wday','timestamp']].groupby(['file_id','tm_wday']).count().reset_index()
april_cnt.columns = ['file_id','tm_wday','frequency']
april_cnt = april_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_wday',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_cnt = may[['file_id','tm_wday','timestamp']].groupby(['file_id','tm_wday']).count().reset_index()
may_cnt.columns = ['file_id','tm_wday','frequency']
may_cnt = may_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_wday',aggfunc='sum').reset_index()

del may
gc.collect();

data = pd.merge(march_cnt,april_cnt,how='outer',on='file_id')
data = pd.merge(data,may_cnt,how='outer',on='file_id')

data = data.fillna(0)
data.columns = [str(i) for i in data.columns]
week_cnt_df = pd.DataFrame(data.file_id.unique())
week_cnt_df.columns = ['file_id']
for i in range(7):
    week_cnt_df[i]=0
    for feat in list(data.columns):
        if str(i) in feat:
            week_cnt_df[i]+=data[feat]

week_cnt_df.to_csv('week_cnt.csv',index=False)

# hour_cnt

march = pd.read_hdf('march.h5', 'df')
march_cnt = march[['file_id','tm_hour','timestamp']].groupby(['file_id','tm_hour']).count().reset_index()
march_cnt.columns = ['file_id','tm_hour','frequency']
march_cnt = march_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_hour',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_cnt = april[['file_id','tm_hour','timestamp']].groupby(['file_id','tm_hour']).count().reset_index()
april_cnt.columns = ['file_id','tm_hour','frequency']
april_cnt = april_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_hour',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_cnt = may[['file_id','tm_hour','timestamp']].groupby(['file_id','tm_hour']).count().reset_index()
may_cnt.columns = ['file_id','tm_hour','frequency']
may_cnt = may_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_hour',aggfunc='sum').reset_index()

del may
gc.collect();


data = pd.concat([march_cnt,april_cnt],axis=0)
data = pd.concat([data,may_cnt],axis=0)

data = data.groupby('file_id').sum().reset_index()
data.columns = ['file_id'] + ['h'+str(i) for i in data.columns.tolist() if i is not 'file_id']

data.fillna(0,inplace=True)

data['sum'] = data.iloc[:,1:].sum(axis=1)
for i in range(24):
    i=i+1
    data.iloc[:,i] = (data.iloc[:,i]  / data['sum'])*100

data.drop('sum',axis=1,inplace=True)
data.head(3)

data.to_csv('hour_cnt.csv',index=False)

# Product

march = pd.read_hdf('march.h5', 'df')
march_prod = march[['file_id','product_id','timestamp']].groupby(['file_id','product_id']).count().reset_index()
march_prod.columns = ['file_id','product_id','frequency']
march_prod = march_prod.pivot_table(values='frequency',index=['file_id'],columns='product_id',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_prod = april[['file_id','product_id','timestamp']].groupby(['file_id','product_id']).count().reset_index()
april_prod.columns = ['file_id','product_id','frequency']
april_prod = april_prod.pivot_table(values='frequency',index=['file_id'],columns='product_id',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_prod = may[['file_id','product_id','timestamp']].groupby(['file_id','product_id']).count().reset_index()
may_prod.columns = ['file_id','product_id','frequency']
may_prod = may_prod.pivot_table(values='frequency',index=['file_id'],columns='product_id',aggfunc='sum').reset_index()

del may
gc.collect();


data = pd.concat([march_prod,april_prod],axis=0)
data = pd.concat([data,may_prod],axis=0)
data = data.groupby('file_id').sum().reset_index()


data.head(3)

data ['prod_sum'] = np.sum(data.iloc[:,1:],axis=1)
for i in range(27):
    i=i+1
    data.iloc[:,i] = (data.iloc[:,i]/data['prod_sum'])

data.drop('prod_sum',axis=1,inplace=True)
data.head(3)

data.to_csv('product_cnt.csv',index=False)

# minute_cnt

march = pd.read_hdf('march.h5', 'df')
march_cnt = march[['file_id','tm_min','timestamp']].groupby(['file_id','tm_min']).count().reset_index()
march_cnt.columns = ['file_id','tm_min','frequency']
march_cnt = march_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_min',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_cnt = april[['file_id','tm_min','timestamp']].groupby(['file_id','tm_min']).count().reset_index()
april_cnt.columns = ['file_id','tm_min','frequency']
april_cnt = april_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_min',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_cnt = may[['file_id','tm_min','timestamp']].groupby(['file_id','tm_min']).count().reset_index()
may_cnt.columns = ['file_id','tm_min','frequency']
may_cnt = may_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_min',aggfunc='sum').reset_index()

del may
gc.collect();

data = pd.concat([march_cnt,april_cnt],axis=0)
data = pd.concat([data,may_cnt],axis=0)
data = data.groupby('file_id').sum().reset_index()

data.head(3)

data.columns = ['file_id'] + ['min'+str(i) for i in data.columns.tolist() if i is not 'file_id']

data.fillna(0,inplace=True)

data['sum'] = data.iloc[:,1:].sum(axis=1)
for i in range(60):
    i=i+1
    data.iloc[:,i] = (data.iloc[:,i]  / data['sum'])*100

data.drop('sum',axis=1,inplace=True)
data.head(3)

data.to_csv('minute_cnt.csv',index=False)

# sec_cnt

march = pd.read_hdf('march.h5', 'df')
march_cnt = march[['file_id','tm_sec','timestamp']].groupby(['file_id','tm_sec']).count().reset_index()
march_cnt.columns = ['file_id','tm_sec','frequency']
march_cnt = march_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_sec',aggfunc='sum').reset_index()
del march
gc.collect();

april = pd.read_hdf('april.h5', 'df')
april_cnt = april[['file_id','tm_sec','timestamp']].groupby(['file_id','tm_sec']).count().reset_index()
april_cnt.columns = ['file_id','tm_sec','frequency']
april_cnt = april_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_sec',aggfunc='sum').reset_index()

del april
gc.collect();

may = pd.read_hdf('may.h5', 'df')
may_cnt = may[['file_id','tm_sec','timestamp']].groupby(['file_id','tm_sec']).count().reset_index()
may_cnt.columns = ['file_id','tm_sec','frequency']
may_cnt = may_cnt.pivot_table(values='frequency',index=['file_id'],columns='tm_sec',aggfunc='sum').reset_index()

del may
gc.collect();

data = pd.concat([march_cnt,april_cnt],axis=0)
data = pd.concat([data,may_cnt],axis=0)
data = data.groupby('file_id').sum().reset_index()

data.head(3)

data.columns = ['file_id'] + ['sec'+str(i) for i in data.columns.tolist() if i is not 'file_id']

data.fillna(0,inplace=True)

data['sum'] = data.iloc[:,1:].sum(axis=1)
for i in range(60):
    i=i+1
    data.iloc[:,i] = (data.iloc[:,i]  / data['sum'])*100

data.drop('sum',axis=1,inplace=True)
data.head(3)

data.to_csv('sec_cnt.csv',index=False)
